fix(auth): make guard implement `CanActivate`.
Use case: the user is not yet logged in and it tries to access the `/users` route, which is protected. As a resuat, the navigation won't be successful.
If the user logged in and then logs out, when `users` is accessed, apparently `CanActivate` is needed because the module is already loaded.